fix: 중복 파일 삭제
